moved to using a main.c
