+++ build.sh
@@ -1,2 +1,2 @@
 #!/bin/bash -ve
-gcc -pthread -O3 -D_GNU_SOURCE -o kakapo kakapo.c session.c util.c sockbuf.c
+gcc -pthread -O3 -D_GNU_SOURCE -o kakapo main.c session.c util.c sockbuf.c
